Elimino archivos deprecados como condicionType y lo relacionado a Beneficio que cambio segun lo hablado en la ultima reunion. Modifico los modelos de curso, inscripcion, pago, rol, usuario y tipo de pago. Segun lo conversado en la ultima reunion (cursos con tipo alquiler y comision, sistema de puntos, beneficios aplicados de forma local en el curso, logica de cobro con recargo por demora en el pago de la cuota). Migro el tipado de Double a BigDecimal para el manejo de dinero. Agrego el rol Oficina hablado en la anteultima reunion.

+++ src/main/kotlin/com/estonianport/centro_sis/model/TipoPago.kt
@@ -5,6 +5,7 @@ import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
 import jakarta.persistence.EnumType
 import jakarta.persistence.Enumerated
+import java.math.BigDecimal
 
 @Embeddable
 data class TipoPago(
@@ -12,5 +13,5 @@ data class TipoPago(
     val tipoPago: PagoType,
 
     @Column(nullable = false)
-    val monto: Double
+    val monto: BigDecimal
 )
